0823 - mention_detail 뷰추가

+++ src/main/java/com/megait/nocoronazone/domain/DetailSafetyIndex.java
@@ -3,8 +3,8 @@ package com.megait.nocoronazone.domain;
 import lombok.*;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+
 @Entity
 @Getter
 @Setter
@@ -17,5 +17,5 @@ public class DetailSafetyIndex {
     @Id
     private String district;
 
-    private double index;
+    private int index;
 }
